chore: rename parser dir to bl
